requirements.txt
appnope==0.1.0
awsebcli==3.14.6
Babel==2.6.0
backcall==0.1.0
blessed==1.15.0
boto3==1.9.39
botocore==1.12.39
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
Django==2.1.2
django-extensions==2.1.3
django-filter==2.0.0
django-nested-admin==3.1.2
django-storages==1.7.1
djangorestframework==3.9.0
docker==3.5.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ez-setup==0.9
idna==2.6
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jmespath==0.9.3
jsonschema==2.6.0
Markdown==3.0.1
parso==0.3.1
pathspec==0.5.5
pexpect==4.6.0
phonenumbers==8.10.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
python-monkey-business==1.0.0
pytz==2018.7
PyYAML==3.13
requests==2.18.4
s3transfer==0.1.13
selenium==3.141.0
semantic-version==2.5.0
sentry-sdk==0.5.5
Serializer==0.2.1
simplejson==3.16.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.54.0
